+++ www/go/core/views/extjs3/tools/SystemSettingsTools.js
@@ -39,6 +39,7 @@ go.tools.SystemSettingsTools = Ext.extend(Ext.Panel,{
 			[t('System check'),'', 'install/gotest.php'],
 			[t('Database check'),'', GO.url('maintenance/checkDatabase') ],
 			[t('Update search index'),'', GO.url('maintenance/buildSearchCache') ],
+			[t('Update search index (Complete rebuild)'),'', GO.url('maintenance/buildSearchCache', {"reset" : 1}) ],
 			[t('Remove duplicate contacts and events'),'', GO.url('maintenance/removeDuplicates') ]
 		];
 		if(go.Modules.isAvailable(null,'files')) {