+++ www/go/core/orm/Entity.php
@@ -193,6 +193,28 @@ abstract class Entity extends Property {
 		
 		return $query;
 	}
+
+	/**
+	 * Find entities linked to the given entity
+	 *
+	 * @param $entity
+	 * @param array $properties
+	 * @param bool $readOnly
+	 * @return Query|static[]
+	 * @throws Exception
+	 */
+	public static function findByLink($entity, $properties = [], $readOnly = false) {
+		$query = static::find($properties, $readOnly);
+		$query->join(
+			'core_link',
+			'l',
+			$query->getTableAlias() . '.id = l.toId and l.toEntityTypeId = '.static::entityType()->getId())
+
+			->andWhere('fromEntityTypeId = '. $entity::entityType()->getId())
+			->andWhere('fromId', '=', $entity->id);
+
+		return $query;
+	}
 	
 //	
 //	public function getId() {		
@@ -1148,6 +1170,35 @@ abstract class Entity extends Property {
 		}
 
 		return $refs;
+	}
+
+	/**
+	 * A title for this entity used in search cache and logging for example.
+	 *
+	 * @return string
+	 */
+	public function title() {
+		if(property_exists($this,'name')) {
+			return $this->name;
+		}
+
+		if(property_exists($this,'title')) {
+			return $this->title;
+		}
+
+		if(property_exists($this,'subject')) {
+			return $this->subject;
+		}
+
+		if(property_exists($this,'description')) {
+			return $this->description;
+		}
+
+		if(property_exists($this,'displayName')) {
+			return $this->displayName;
+		}
+
+		return static::class;
 	}
 	
 }