+++ www/go/modules/community/ldapauthenticator/model/Server.php
@@ -41,6 +41,8 @@ class Server extends Entity {
 	public $syncGroups = false;
 	public $syncGroupsQuery;
 
+	public $imapUseEmailForUsername = false;
+
 	
 	
 	/**