+++ www/views/Extjs3/lang.php
@@ -9,7 +9,6 @@ if(isset($_GET['lang'])) {
 	go()->getLanguage()->setLanguage($_GET['lang']);
 }
 
-header('Content-Type: application/javascript; charset=utf8');
 $webclient = Extjs3::get();
 $webclient->getLanguageJS()->output(true, true, [
 	"Expires" => (new DateTime("1 year"))->format("D, j M Y H:i:s")