+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Billing: Field 'invoice_no' is now a varchar
 - Core: Fixed 500 error upon rebuilding module cache
 - Core: Fixed custom fields saving new select box.
 - Core: Custom fields, generate database name as per MariaDB naming conventions.