+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     	- Strict-Transport-Security: max-age=31536000
         - X-XSS-Protection: 1;mode=block
 - Core: Use relative URL's in webclient
+- Email: reload grid and keep position but don't select next mail
 
 30-06-2020 6.4.155
 - Email: Problem with empty emails