+++ CHANGELOG.md
@@ -1,5 +1,22 @@
+- Billing: Field 'invoice_no' is now a varchar
+- Core: Fixed 500 error upon rebuilding module cache
+- Core: Fixed custom fields saving new select box.
+- Core: Custom fields, generate database name as per MariaDB naming conventions.
+- Address book: Update search cache of employees when changing company name
+- LDAP Auth: Added option to login to IMAP server with email instead of username 
+- Core: Added complete rebuild search cache option.
+- Core: Fixed broken link display in e-mail
+- Core: Set security headers: 
+    	- X-Frame-Options: SAMEORIGIN
+        - Content-Security-Policy
+    	- X-Content-Type-Options: nosniff
+    	- Strict-Transport-Security: max-age=31536000
+        - X-XSS-Protection: 1;mode=block
+- Core: Use relative URL's in webclient
+- Email: reload grid and keep position but don't select next mail
+
 30-06-2020 6.4.155
-- Emnail: Problem with empty emails
+- Email: Problem with empty emails
 
 30-06-2020 6.4.154
 - Custom fields: Changing multi select options could destroy all values of the field